# contracts/source-connector-v1.openapi.yaml
openapi: 3.1.0
info:
  title: AutoData Source Connector
  version: 1.0.0
  description: Provider-neutral read-only contract implemented independently by Bankone and Banktwo.
servers:
  - url: https://bankone.cars.tk
  - url: https://banktwo.cars.tk
paths:
  /v1/capabilities:
    get:
      operationId: capabilities
      responses:
        '200':
          description: Supported source operations.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CapabilitiesResponse' }
  /v1/catalog/{scope}:
    get:
      operationId: readCatalog
      parameters:
        - { name: scope, in: path, required: true, schema: { type: string, enum: [years, makes, models, configurations] } }
        - { name: year, in: query, schema: { type: integer, minimum: 1886, maximum: 2200 } }
        - { name: make, in: query, schema: { type: string, maxLength: 160 } }
        - { name: model, in: query, schema: { type: string, maxLength: 160 } }
        - { name: cursor, in: query, schema: { type: string, maxLength: 512 } }
      responses:
        '200':
          description: A bounded catalog page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogResponse' }
        default: { $ref: '#/components/responses/Error' }
  /v1/vehicle-resolutions:
    post:
      operationId: resolveVehicle
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VehicleSelector' }
      responses:
        '200':
          description: Zero or more source vehicle candidates.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VehicleResolutionResponse' }
        default: { $ref: '#/components/responses/Error' }
  /v1/vehicles/{opaqueRef}/articles:
    get:
      operationId: listArticles
      parameters:
        - { name: opaqueRef, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 1024 } }
        - { name: cursor, in: query, schema: { type: string, maxLength: 512 } }
      responses:
        '200':
          description: A bounded article page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArticleListResponse' }
        default: { $ref: '#/components/responses/Error' }
  /v1/vehicles/{opaqueRef}/article-search:
    post:
      operationId: searchArticles
      parameters:
        - { name: opaqueRef, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 1024 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              additionalProperties: false
              properties:
                query: { type: string, minLength: 1, maxLength: 512 }
                cursor: { type: string, maxLength: 512 }
      responses:
        '200':
          description: A bounded query result page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArticleListResponse' }
        default: { $ref: '#/components/responses/Error' }
  /v1/resources/{opaqueRef}:
    get:
      operationId: readResource
      parameters:
        - { name: opaqueRef, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 2048 } }
      responses:
        '200':
          description: Raw source bytes with provenance headers.
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
            X-Provider: { required: true, schema: { type: string, enum: [bankone, banktwo] } }
            X-Source-Revision: { required: true, schema: { type: string, maxLength: 256 } }
            X-Fetched-At: { required: true, schema: { type: string, format: date-time } }
            X-Source-Locator: { required: true, schema: { type: string, maxLength: 2048 } }
            X-Source-Sha256: { required: true, schema: { type: string, pattern: '^[a-f0-9]{64}$' } }
            X-Source-Media-Type: { required: true, schema: { type: string, maxLength: 128 } }
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/TextResource' }
                  - { $ref: '#/components/schemas/BinaryResource' }
            application/octet-stream:
              schema: { type: string, format: binary }
        default: { $ref: '#/components/responses/Error' }
  /healthz:
    get:
      operationId: health
      responses: { '200': { description: Healthy } }
  /readyz:
    get:
      operationId: readiness
      responses: { '200': { description: Ready }, '503': { description: Not ready } }
components:
  headers:
    RequestId:
      description: Correlation identifier generated by the bank.
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Stable source connector error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    SourceEnvelope:
      type: object
      required: [request_id, provider, source_revision, fetched_at]
      properties:
        request_id: { type: string, format: uuid }
        provider: { type: string, enum: [bankone, banktwo] }
        source_revision: { type: string, minLength: 1, maxLength: 256 }
        fetched_at: { type: string, format: date-time }
        source_locator: { type: string, maxLength: 2048 }
    Page:
      type: object
      required: [complete]
      oneOf:
        - properties: { complete: { const: true } }
        - required: [next_cursor]
          properties: { complete: { const: false }, next_cursor: { type: string, maxLength: 512 } }
      properties:
        complete: { type: boolean }
        next_cursor: { type: string, maxLength: 512 }
    CatalogItem:
      type: object
      required: [opaque_ref, label]
      additionalProperties: false
      properties:
        opaque_ref: { type: string, minLength: 1, maxLength: 1024 }
        label: { type: string, minLength: 1, maxLength: 512 }
        year: { type: integer }
        make: { type: string, maxLength: 160 }
        model: { type: string, maxLength: 160 }
        configuration: { type: string, maxLength: 512 }
        source_label: { type: string, maxLength: 512 }
    CatalogResponse:
      allOf:
        - { $ref: '#/components/schemas/SourceEnvelope' }
        - { $ref: '#/components/schemas/Page' }
        - type: object
          required: [scope, items]
          additionalProperties: false
          properties:
            request_id: {}
            provider: {}
            source_revision: {}
            fetched_at: {}
            source_locator: {}
            complete: {}
            next_cursor: {}
            scope: { type: string, enum: [years, makes, models, configurations] }
            items: { type: array, maxItems: 1000, items: { $ref: '#/components/schemas/CatalogItem' } }
    VehicleSelector:
      type: object
      required: [year, make, model]
      additionalProperties: false
      properties:
        year: { type: integer, minimum: 1886, maximum: 2200 }
        make: { type: string, minLength: 1, maxLength: 160 }
        model: { type: string, minLength: 1, maxLength: 160 }
        configuration: { type: string, maxLength: 512 }
        region: { type: string, maxLength: 64 }
        vin: { type: string, maxLength: 32 }
    VehicleCandidate:
      type: object
      required: [opaque_ref, label, confidence]
      additionalProperties: false
      properties:
        opaque_ref: { type: string, minLength: 1, maxLength: 1024 }
        label: { type: string, minLength: 1, maxLength: 512 }
        confidence: { type: number, minimum: 0, maximum: 1 }
        evidence: { type: array, maxItems: 32, items: { type: string, maxLength: 512 } }
    VehicleResolutionResponse:
      allOf:
        - { $ref: '#/components/schemas/SourceEnvelope' }
        - type: object
          required: [selector, candidates]
          additionalProperties: false
          properties:
            request_id: {}
            provider: {}
            source_revision: {}
            fetched_at: {}
            source_locator: {}
            selector: { $ref: '#/components/schemas/VehicleSelector' }
            candidates: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/VehicleCandidate' } }
    Article:
      type: object
      required: [opaque_ref, title]
      additionalProperties: false
      properties:
        opaque_ref: { type: string, minLength: 1, maxLength: 2048 }
        title: { type: string, minLength: 1, maxLength: 1000 }
        category: { type: string, maxLength: 256 }
        component: { type: string, maxLength: 256 }
        labor_resource_ref: { type: string, maxLength: 2048 }
        resource_ref: { type: string, maxLength: 2048 }
    ArticleListResponse:
      allOf:
        - { $ref: '#/components/schemas/SourceEnvelope' }
        - { $ref: '#/components/schemas/Page' }
        - type: object
          required: [articles]
          additionalProperties: false
          properties:
            request_id: {}
            provider: {}
            source_revision: {}
            fetched_at: {}
            source_locator: {}
            complete: {}
            next_cursor: {}
            articles: { type: array, maxItems: 1000, items: { $ref: '#/components/schemas/Article' } }
    TextResource:
      allOf:
        - { $ref: '#/components/schemas/SourceEnvelope' }
        - type: object
          required: [kind, media_type, content, sha256]
          additionalProperties: false
          properties:
            request_id: {}
            provider: {}
            source_revision: {}
            fetched_at: {}
            source_locator: {}
            kind: { type: string, enum: [article, labor, text] }
            media_type: { type: string, maxLength: 128 }
            content: { type: string, maxLength: 10000000 }
            sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
    BinaryResource:
      allOf:
        - { $ref: '#/components/schemas/SourceEnvelope' }
        - type: object
          required: [kind, media_type, content_base64, sha256]
          additionalProperties: false
          properties:
            request_id: {}
            provider: {}
            source_revision: {}
            fetched_at: {}
            source_locator: {}
            kind: { type: string, enum: [asset, binary] }
            media_type: { type: string, maxLength: 128 }
            content_base64: { type: string, maxLength: 20000000 }
            sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
    ErrorResponse:
      type: object
      required: [request_id, error]
      additionalProperties: false
      properties:
        request_id: { type: string, format: uuid }
        error:
          type: object
          required: [code, message, retryable]
          additionalProperties: false
          properties:
            code: { type: string, enum: [INVALID_INPUT, NOT_FOUND, AMBIGUOUS, UNAUTHORIZED, RATE_LIMITED, UPSTREAM_UNAVAILABLE, INVALID_UPSTREAM_RESPONSE] }
            message: { type: string, minLength: 1, maxLength: 1000 }
            retryable: { type: boolean }
            details:
              type: object
              additionalProperties: true
              propertyNames:
                pattern: '^(?!(authorization|cookie|password|secret|token|database|db_id)$)[A-Za-z0-9_.-]+$'
    CapabilitiesResponse:
      allOf:
        - { $ref: '#/components/schemas/SourceEnvelope' }
        - type: object
          required: [capabilities]
          additionalProperties: false
          properties:
            request_id: {}
            provider: {}
            source_revision: {}
            fetched_at: {}
            source_locator: {}
            capabilities: { type: array, uniqueItems: true, items: { type: string, enum: [catalog, vehicle_resolution, article_list, article_search, resource_read] } }